vdk-control-cli: fix circular import dependecy
There is a circular import issue between the
vdk.internal.control.configuration.vdk_config and
vdk.internal.control.command_groups.version_group.version modules.
Specifically, vdk_config is importing version and version is importing
cli_utils, which is importing vdk_config. This creates a cycle that
cannot be resolved, which is causing the ImportError under certain
circumstances.

To fix this issue, I need to refactor the code to remove the circular
import. I moved the methods for determining verison in version utils
which is used by both vdk_config and version_group.version modules.
