+++ projects/vdk-control-cli/src/vdk/internal/control/command_groups/version_group/version.py
@@ -1,31 +1,8 @@
 # Copyright 2021-2023 VMware, Inc.
 # SPDX-License-Identifier: Apache-2.0
 import click
-from pkg_resources import DistributionNotFound
-from pkg_resources import get_distribution
-
-# https://packaging.python.org/guides/single-sourcing-package-version/#
-
-try:
-    # Change here if project is renamed and does not equal the setuptools metadata.name
-    dist_name = "vdk-control-cli"
-    __version__ = get_distribution(dist_name).version
-except DistributionNotFound:  # pragma: no cover
-    __version__ = "unknown"
-
-
-def build_details():
-    try:
-        from vdk.internal.control import vdk_control_build_info
-
-        build = [
-            f"{key}={value}"
-            for key, value in vdk_control_build_info.__dict__.items()
-            if not key.startswith("_")
-        ]
-        return ", ".join(build)
-    except:
-        return ""
+from vdk.internal.control.utils.version_utils import __version__
+from vdk.internal.control.utils.version_utils import build_details
 
 
 @click.command(help="Prints the version of the client")
